+++ js/utils.js
@@ -1,0 +1,1 @@
+export const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
